Refactor code so that all code related to output transofmrations is moved to _module_output_transformation.py
